fix: fix issues regarding pipeline
