+++ src/app/features/private/shared/components/group-task-from/group-task-form.component.ts
@@ -164,7 +164,8 @@ export class GroupTaskFormComponent {
         0,
         0,
       );
-      let editGroupTaskRequest: EditGroupTaskDto = {
+
+      const editGroupTaskRequest: EditGroupTaskDto = {
         title: formValue.title,
         description: formValue.description,
         startTime: formValue.startTime!.toISOString(),
