+++ src/app/features/private/shared/components/group-tasks-list/group-tasks-list.component.ts
@@ -10,7 +10,10 @@ import {
 import { GroupTaskFormComponent } from '../group-task-from/group-task-form.component';
 import { NzModalModule } from 'ng-zorro-antd/modal';
 import { GroupTaskApiService } from '../../../../shared/services/group-task-api/group-task-api.service';
-import { GroupTask } from '../../../../shared/models/group/group-task.model';
+import {
+  GetGroupTaskRequest,
+  GroupTask,
+} from '../../../../shared/models/group/group-task.model';
 import { Group } from '../../../../shared/models/group/group.model';
 import { GroupPreviewMode } from '../../../../shared/models/group/group-preview-mode';
 import { CommonModule } from '@angular/common';
@@ -52,7 +55,7 @@ export class GroupTasksListComponent implements OnInit {
   protected loading = signal<boolean>(true);
   protected hasMore = signal<boolean>(true);
   protected tasksList = signal<GroupTask[]>([]);
-  protected tasksRequestParams = {
+  protected tasksRequestParams: GetGroupTaskRequest = {
     isAccepted: false,
     isDeclined: false,
     page: 0,
@@ -90,9 +93,9 @@ export class GroupTasksListComponent implements OnInit {
   loadMoreTasks(): void {
     if (!this.hasMore() || this.loadingMore()) return;
     this.loadingMore.set(true);
-    let nexPage = this.tasksRequestParams.page + 1;
+    const nexPage = this.tasksRequestParams.page + 1;
 
-    let nexPageParams = {
+    const nexPageParams = {
       size: this.tasksRequestParams.size,
       page: nexPage,
       isAccepted: this.tasksRequestParams.isAccepted,
