+++ src/app/features/private/shared/components/input-search/input-search.component.ts
@@ -14,7 +14,7 @@ import { CommonModule } from '@angular/common';
 import { GroupApiService } from '../../../../shared/services/groups-api/group-api.service';
 import { HttpClient } from '@angular/common/http';
 import { NzSelectModule } from 'ng-zorro-antd/select';
-import { GroupFilterParams } from '../../../../shared/models/group/group-filter-params.model';
+import { GroupFilterParams } from '../../../../shared/models/group/group.model';
 import { GroupType } from '../../../../shared/models/group/group-type.model';
 
 @Component({
