--- src/app/features/shared/models/group/group-filter-params.model.ts
@@ -1,8 +1,0 @@
-export interface GroupFilterParams {
-  joinCode?: string;
-  groupType?: number;
-  groupName?: string;
-  isForLoggedUser?: boolean;
-  page: number;
-  size: number;
-}
